requirements.txt
flask
gunicorn
social-auth-core
social-auth-app-flask
social-auth-app-flask-sqlalchemy
social-auth-storage-sqlalchemy
flask-login
python-jose
pycrypto
pyjwt
sqlalchemy
requests
pyyaml
psycopg2-binary
wtforms
paginate
paginate_sqlalchemy
google-cloud-storage
bs4
lxml
google_auth_httplib2
google_auth_oauthlib==0.3.0
google-api-python-client
python-datauri
openpyxl
ua-parser
user-agents
